cambios a vista de errores
